backend/requirements.txt
Django~=4.2.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
channels~=4.0.0
googlemaps~=4.10.0
psycopg2-binary~=2.9.9
whitenoise~=6.5.0
python-dotenv~=1.0.0
setuptools~=70.0.0